# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "WebNewsAPI",
        "contact": {}
    },
    "paths": {
        "/article": {
            "get": {
                "description": "Show a Article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Show Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.GetArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Article for the app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Create Articles",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.CreateArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Delete Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.DeleteArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a Article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Update Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "article data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.UpdateArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "description": "List all articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "List Articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ListArticleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/articleshandlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listnews": {
            "get": {
                "description": "List all news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "List news",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ListNewsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "description": "Show a news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Show news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.GetNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a news for the app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Create News",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newsHandler.CreateNewsObject"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.CreateNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Delete news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.DeleteNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Update news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "News data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newsHandler.UpdateNewsObject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.UpdateNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "articleshandlers.CreateArticleRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.CreateArticleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.ArticlesResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.DeleteArticleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.ArticlesResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.GetArticleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.ArticlesResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.ListArticleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ArticlesResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.UpdateArticleRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "articleshandlers.UpdateArticleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.ArticlesResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.CreateNewsObject": {
            "type": "object",
            "properties": {
                "firstContent": {
                    "type": "string"
                },
                "secondContent": {
                    "type": "string"
                },
                "thirdContent": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "newsHandler.CreateNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.DeleteNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.GetNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.ListNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.NewsResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.UpdateNewsObject": {
            "type": "object",
            "properties": {
                "firstContent": {
                    "type": "string"
                },
                "secondContent": {
                    "type": "string"
                },
                "thirdContent": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "newsHandler.UpdateNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "schema.ArticlesResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "schema.NewsResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "firstContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "secondContent": {
                    "type": "string"
                },
                "thirdContent": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}